feature(Optimizer): changed get_candidates() api so candidate evaluation only steps when an array with length of zero is returned
